requirements.txt
azure-functions
requests==2.22.0